[redis] Publish QoS value from `Filter-Id`
This commit allows passing QoS profiles to accel-ppp via `Filter-Id`
Attributes in radius Access-Accept messages. The value in `Filter-Id` is
parsed when the Access-Accept message is received from radius and stored
in the newly introduced `qos_val` member of the ap_session struct to
make the QoS profile available across different events. When the
`qos_val` is set, it will be inserted into the JSON messages published
on the redis channel.